refactor(simulator): extract IPC subsystem into simulator/ipc.rs (Step 17)
Move the mailbox/semaphore/named-event waiter structs (SemGetWaiter,
MailboxGetWaiter, MailboxPutWaiter, EventWaiter, InstanceEventWaiter) and the
waiter-driver methods (make_event_waiter, make_event_waiter_kind,
drain_triggered_event_waiters, fire_instance_event, wake_semaphore_waiters,
admit_mailbox_put_waiter) into a dedicated simulator/ipc.rs submodule.
mod.rs shrinks by ~330 lines. 1935 tests pass.
